+++ avocado/utils/external/spark.py
