+++ examples/tests/modify_variable.py
@@ -3,9 +3,9 @@
 import os
 import shutil
 
-from avocado import gdb
 from avocado import main
 from avocado import Test
+from avocado.utils import gdb
 from avocado.utils import build
